refactor: location => cmdb
