+++ pkg/plugin/cmdb/location.go
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package location
+package cmdb
 
 import (
 	"github.com/polarismesh/polaris-go/pkg/model"
